@@ -1,29 +1,15 @@
 {
   "name": "gp-boilerplate",
-  "version": "4.0.0",
+  "version": "5.0.0",
   "description": "Grabarz & Partner - Frontend Boilerplate",
   "main": "none",
   "engines": {
-    "node": ">=8"
+    "node": ">=10"
   },
-  "scripts": {
-    "heroku": "NODE_ENV=production node node_modules/gulp/bin/gulp run --serverConfig=./env/config/server.json --gulpFileMap=./env/config/gulp-file-map.json --server=gp-boilerplate-webserver",
-    "build": "NODE_ENV=build node node_modules/gulp/bin/gulp build --serverConfig=/env/config/hapi/config --gulpTaskConfig=/env/config/",
-    "dev": "NODE_ENV=development node node_modules/gulp/bin/gulp run --serverConfig=/env/config/hapi/config --gulpTaskConfig=/env/config/ --server=gp-boilerplate-webserver",
-    "dev-ip": "NODE_ENV=development node node_modules/gulp/bin/gulp run --serverConfig=/env/config/hapi/config --gulpTaskConfig=/env/config/ --server=gp-boilerplate-webserver --ip=true",
-    "prod": "NODE_ENV=production node node_modules/gulp/bin/gulp run --serverConfig=/env/config/hapi/config --gulpTaskConfig=/env/config/ --server=gp-boilerplate-webserver",
-    "server": "NODE_ENV=development node node_modules/gulp/bin/gulp default --serverConfig=/env/config/hapi/config --server=gp-boilerplate-webserver",
-    "test": "npm run build",
-    "template": "yo gp-boilerplate:hbs-template",
-    "template-package": "yo gp-boilerplate:hbs-package-template",
-    "create-package": "yo gp-boilerplate:package",
-    "image-opt": "yo gp-boilerplate:image-opt",
-    "svg-opt": "yo gp-boilerplate:svg-opt",
-    "test-task-grid": "NODE_ENV=build node node_modules/gulp/bin/gulp grid --serverConfig=/env/config/hapi/config --gulpTaskConfig=/env/config/",
-    "test-task-svg-symbols": "NODE_ENV=build node node_modules/gulp/bin/gulp svg-symbols --serverConfig=/env/config/hapi/config --gulpTaskConfig=/env/config/",
-    "test-task-svgo": "NODE_ENV=build node node_modules/gulp/bin/gulp svgo --serverConfig=/env/config/hapi/config --gulpTaskConfig=/env/config/",
-    "test-task-handlebars": "NODE_ENV=development node node_modules/gulp/bin/gulp handlebars --serverConfig=/env/config/hapi/config --gulpTaskConfig=/env/config/ --server=gp-boilerplate-webserver"
+  "bugs": {
+    "url": "https://github.com/GrabarzUndPartner/gp-boilerplate/issues"
   },
+  "homepage": "https://github.com/GrabarzUndPartner/gp-boilerplate#readme",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/GrabarzUndPartner/gp-boilerplate.git"
