@@ -37,8 +37,8 @@ module.exports = {
     },
     // options for formating the statistics
 
-    reporter: null,
-    // Provide a custom reporter to change the way how logs are shown.
+    // reporter: null,
+    // // Provide a custom reporter to change the way how logs are shown.
 
     serverSideRender: false
     // Turn off the server-side rendering mode. See Server-Side Rendering part for more info.
