@@ -1,14 +1,11 @@
-'use strict';
 
-// IE 11
-import 'element-closest';
+document.documentElement.classList.remove('no-js');
 
 import 'script-loader!fg-loadcss';
 import 'script-loader!fg-loadcss/src/cssrelpreload';
-import 'gp-module-polyfills';
 import './embed/fontfaceObserver';
-import './embed/prefix';
 
-// SVG Reference Polyfill
-import svg4everybody from 'svg4everybody';
-svg4everybody();
+import objectFitImages from 'object-fit-images';
+Array.from(document.querySelectorAll('picture')).forEach(function (el) {
+    objectFitImages(el);
+});
